Restaurant changes completed

+++ FoodOrderingApp-service/src/main/java/com/upgrad/FoodOrderingApp/service/businness/StateBusinessService.java
@@ -14,7 +14,9 @@ public class StateBusinessService {
     private StateDao stateDao;
 
     @Transactional
-    public StateEntity getStateById (final Long stateId) {
+    // A Method which takes the stateId as parameter for  getStateById endpoint
+    public StateEntity getStateById (final Integer stateId) {
+
         return stateDao.getStateById(stateId);
     }
 }
